// Front-End/src/app/models/notification/notification.ts
export class Notification {
    companyId: number | null = null;
    resaleId: number | null = null;
    id: number | null = null;
    origUserId: number | null = null;
    origUserName: string = '';
    origDate: Date | string = '';
    origRoleId: number | null = null;
    origRoleDesc: string = '';
    origNotificationMenu: string = '';
    origId: string = '';
    header: string = '';
    message1: string = '';
    message2: string = '';
    message3: string = '';
}